# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Amazon",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/productos": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Get list of productos",
                "description": "Returns list of productos",
                "operationId": "496dc08306524d77f18d124d1d6838aa",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/producto/": {
            "post": {
                "tags": [
                    "Productos"
                ],
                "description": "Add productos",
                "operationId": "fbc588df3d7222c4c0441e48e0bc44e1",
                "requestBody": {
                    "description": "Pass producto info",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "precio",
                                    "marca",
                                    "descripcion",
                                    "nombre"
                                ],
                                "properties": {
                                    "precio": {
                                        "type": "integer",
                                        "format": "text",
                                        "example": ""
                                    },
                                    "marca": {
                                        "type": "string",
                                        "format": "text",
                                        "example": ""
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "format": "text",
                                        "example": ""
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "format": "text",
                                        "example": ""
                                    },
                                    "image": {
                                        "type": "multipart/form-data",
                                        "format": "file",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong data introduced"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/producto/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Get a producto",
                "description": "Returns one producto",
                "operationId": "af41299eb66ce2d7e6a7c81112895d23",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Update existing producto",
                "description": "Returns updated producto data",
                "operationId": "e295410d9e6c7e46dd5cb53161d676ab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "type": "string",
                                        "format": "text",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Delete existing producto",
                "description": "Deletes a record and returns no content",
                "operationId": "26f9b1a91c32db076a712a31ed7ddd2c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}